Ignore superflat and debug worlds

+++ src/main/java/uk/debb/vanilla_disable/mixin/worldgen/biomes/MixinPlayerManager.java
@@ -12,7 +12,8 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.text.TranslatableText;
 import net.minecraft.util.WorldSavePath;
 import net.minecraft.world.GameRules;
-
+import net.minecraft.world.gen.chunk.DebugChunkGenerator;
+import net.minecraft.world.gen.chunk.FlatChunkGenerator;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
@@ -87,10 +88,10 @@ public abstract class MixinPlayerManager {
     @Inject(method = "onPlayerConnect", at = @At("RETURN"), cancellable = true)
     private void onPlayerConnect(ClientConnection connection, ServerPlayerEntity player, CallbackInfo ci) throws InterruptedException {
         GameRules gameRules = RegisterGamerules.getServer().getGameRules();
-        if ((gameRules.getBoolean(RegisterGamerules.REMOVE_OVERWORLD_BIOMES) ||
-             gameRules.getBoolean(RegisterGamerules.REMOVE_NETHER_BIOMES) ||
-             gameRules.getBoolean(RegisterGamerules.REMOVE_END_BIOMES)) &&
-             player.getWorld().getTimeOfDay() < 100) {
+        if (!(RegisterGamerules.getServer().getOverworld().getChunkManager().getChunkGenerator() instanceof FlatChunkGenerator) &&
+            !(RegisterGamerules.getServer().getOverworld().getChunkManager().getChunkGenerator() instanceof DebugChunkGenerator) &&
+            (gameRules.getBoolean(RegisterGamerules.REMOVE_OVERWORLD_BIOMES) || gameRules.getBoolean(RegisterGamerules.REMOVE_NETHER_BIOMES) ||
+             gameRules.getBoolean(RegisterGamerules.REMOVE_END_BIOMES)) && player.getWorld().getTimeOfDay() < 100) {
             deleteRegionFilesAndClose();
         }
     }
